Refactor: move types from AppContext to /types
